v0.0.3
Update Electron version: 11 -> 18
Update building setup: Webpack -> Vite
Update Vue: 2.6 -> 3.2
