+++ package.json
@@ -1,103 +1,43 @@
 {
   "name": "sight",
-  "productName": "sight",
-  "version": "1.0.0",
-  "description": "Fast image viewer",
-  "main": ".webpack/main",
-  "scripts": {
-    "start": "electron-forge start",
-    "package": "electron-forge package",
-    "make": "electron-forge make",
-    "publish": "electron-forge publish",
-    "lint": "eslint --ext .ts .",
-    "lint:fix": "eslint --fix --ext .ts ."
-  },
-  "keywords": [],
+  "version": "0.0.3",
+  "main": "dist/main/index.cjs",
   "author": {
     "name": "Val Saven",
     "email": "val.saven@gmail.com"
   },
   "license": "GPL-3.0",
-  "config": {
-    "forge": {
-      "packagerConfig": {},
-      "makers": [
-        {
-          "name": "@electron-forge/maker-squirrel",
-          "config": {
-            "name": "sight"
-          }
-        },
-        {
-          "name": "@electron-forge/maker-zip",
-          "platforms": [
-            "darwin"
-          ]
-        },
-        {
-          "name": "@electron-forge/maker-deb",
-          "config": {}
-        },
-        {
-          "name": "@electron-forge/maker-rpm",
-          "config": {}
-        }
-      ],
-      "plugins": [
-        [
-          "@electron-forge/plugin-webpack",
-          {
-            "mainConfig": "./webpack.main.config.js",
-            "renderer": {
-              "config": "./webpack.renderer.config.js",
-              "entryPoints": [
-                {
-                  "html": "./src/index.html",
-                  "js": "./src/renderer.ts",
-                  "name": "main_window"
-                }
-              ]
-            }
-          }
-        ]
-      ]
-    }
+  "scripts": {
+    "dev": "node scripts/watch.mjs",
+    "prebuild": "vue-tsc --noEmit --p packages/renderer/tsconfig.json && node scripts/build.mjs",
+    "build": "electron-builder"
+  },
+  "engines": {
+    "node": ">=17.8.0"
   },
   "devDependencies": {
-    "@electron-forge/cli": "^6.0.0-beta.54",
-    "@electron-forge/maker-deb": "^6.0.0-beta.54",
-    "@electron-forge/maker-rpm": "^6.0.0-beta.54",
-    "@electron-forge/maker-squirrel": "^6.0.0-beta.54",
-    "@electron-forge/maker-zip": "^6.0.0-beta.54",
-    "@electron-forge/plugin-webpack": "^6.0.0-beta.54",
-    "@marshallofsound/webpack-asset-relocator-loader": "^0.5.0",
-    "@typescript-eslint/eslint-plugin": "^4.22.0",
-    "@typescript-eslint/parser": "^4.22.0",
-    "@vue/eslint-config-airbnb": "^5.3.0",
-    "@vue/eslint-config-typescript": "^7.0.0",
-    "css-loader": "^5.2.1",
-    "electron": "^11.3.0",
-    "eslint": "^7.24.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-vue": "^7.9.0",
-    "fork-ts-checker-webpack-plugin": "^6.2.1",
-    "node-loader": "^1.0.3",
-    "string-replace-loader": "^3.0.1",
-    "style-loader": "^2.0.0",
-    "ts-loader": "^8.1.0",
-    "typescript": "^4.2.4",
-    "vue-loader": "^15.9.6",
-    "vue-svg-inline-loader": "^2.1.2",
-    "vue-template-compiler": "^2.6.12"
+    "@typescript-eslint/parser": "5.18.0",
+    "@vitejs/plugin-vue": "2.3.1",
+    "date-fns": "2.28.0",
+    "electron": "18.0.3",
+    "electron-builder": "23.0.3",
+    "eslint": "8.13.0",
+    "eslint-plugin-vue": "8.6.0",
+    "typescript": "4.6.3",
+    "vite": "2.9.1",
+    "vite-plugin-electron": "0.4.0",
+    "vite-plugin-resolve": "2.0.2",
+    "vue": "3.2.31",
+    "vue-tsc": "0.34.2",
+    "vue-virtual-scroller": "2.0.0-alpha.1"
+  },
+  "env": {
+    "VITE_DEV_SERVER_HOST": "127.0.0.1",
+    "VITE_DEV_SERVER_PORT": 3344
   },
   "dependencies": {
-    "date-fns": "^2.20.2",
-    "electron-squirrel-startup": "^1.0.0",
-    "image-size": "^0.9.7",
-    "md5-file": "^5.0.0",
-    "trash": "^7.1.1",
-    "vue": "^2.6.12",
-    "vue-property-decorator": "^9.1.2",
-    "vue-virtual-scroller": "^1.0.10"
+    "image-size": "1.0.1",
+    "md5-file": "5.0.0",
+    "trash": "8.1.0"
   }
 }
