+++ packages/main/vite.config.ts
@@ -1,0 +1,25 @@
+import { builtinModules } from 'module'
+import { defineConfig } from 'vite'
+import pkg from '../../package.json'
+
+export default defineConfig({
+  root: __dirname,
+  build: {
+    outDir: '../../dist/main',
+    lib: {
+      entry: 'index.ts',
+      formats: ['cjs'],
+      fileName: () => '[name].cjs',
+    },
+    minify: process.env./* from mode option */NODE_ENV === 'production',
+    sourcemap: true,
+    rollupOptions: {
+      external: [
+        'electron',
+        ...builtinModules,
+        // @ts-ignore
+        ...Object.keys(pkg.dependencies || {}),
+      ],
+    },
+  },
+})
