+++ packages/preload/vite.config.ts
@@ -1,0 +1,26 @@
+import { builtinModules } from 'module'
+import { defineConfig } from 'vite'
+import pkg from '../../package.json'
+
+export default defineConfig({
+  root: __dirname,
+  build: {
+    outDir: '../../dist/preload',
+    lib: {
+      entry: 'index.ts',
+      formats: ['cjs'],
+      fileName: () => '[name].cjs',
+    },
+    minify: process.env./* from mode option */NODE_ENV === 'production',
+    // https://github.com/caoxiemeihao/electron-vue-vite/issues/61
+    sourcemap: 'inline',
+    rollupOptions: {
+      external: [
+        'electron',
+        ...builtinModules,
+        // @ts-ignore
+        ...Object.keys(pkg.dependencies || {}),
+      ],
+    },
+  },
+})
