+++ packages/renderer/src/global.d.ts
@@ -1,0 +1,11 @@
+
+export { }
+
+declare global {
+  interface Window {
+    // Expose some Api through preload script
+    fs: typeof import('fs')
+    ipcRenderer: import('electron').IpcRenderer
+    removeLoading: () => void
+  }
+}
