+++ scripts/build.mjs
@@ -1,0 +1,5 @@
+import { build } from 'vite'
+
+await build({ configFile: 'packages/main/vite.config.ts' })
+await build({ configFile: 'packages/preload/vite.config.ts' })
+await build({ configFile: 'packages/renderer/vite.config.ts' })
