--- src/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <title>Sight</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <!-- TODO: Remove before release -->
-    <script src="http://localhost:8098"></script>
-  </body>
-</html>
