--- src/index.ts
@@ -1,63 +1,0 @@
-import { app, BrowserWindow, protocol } from 'electron';
-
-declare const MAIN_WINDOW_WEBPACK_ENTRY: any;
-
-// Handle creating/removing shortcuts on Windows when installing/uninstalling.
-if (require('electron-squirrel-startup')) { // eslint-disable-line global-require
-  app.quit();
-}
-
-const createWindow = () => {
-  const protocolName = 'safe-file-protocol';
-
-  protocol.registerFileProtocol(protocolName, (request, callback) => {
-    const url = request.url.replace(`${protocolName}://`, '');
-    try {
-      return callback(decodeURIComponent(url));
-    } catch (error) {
-      // Handle the error as needed
-      console.error(error);
-    }
-  });
-
-  // Create the browser window.
-  const mainWindow = new BrowserWindow({
-    height: 720,
-    width: 1280,
-    webPreferences: {
-      webSecurity: false,
-      nodeIntegration: true,
-    },
-  });
-
-  // and load the index.html of the app.
-  mainWindow.loadURL(MAIN_WINDOW_WEBPACK_ENTRY);
-
-  // Open the DevTools.
-  mainWindow.webContents.openDevTools();
-};
-
-// This method will be called when Electron has finished
-// initialization and is ready to create browser windows.
-// Some APIs can only be used after this event occurs.
-app.on('ready', createWindow);
-
-// Quit when all windows are closed, except on macOS. There, it's common
-// for applications and their menu bar to stay active until the user quits
-// explicitly with Cmd + Q.
-app.on('window-all-closed', () => {
-  if (process.platform !== 'darwin') {
-    app.quit();
-  }
-});
-
-app.on('activate', () => {
-  // On OS X it's common to re-create a window in the app when the
-  // dock icon is clicked and there are no other windows open.
-  if (BrowserWindow.getAllWindows().length === 0) {
-    createWindow();
-  }
-});
-
-// In this file you can include the rest of your app's specific main process
-// code. You can also put them in separate files and import them here.
