--- src/shims-vue.d.ts
@@ -1,5 +1,0 @@
-declare module '*.vue' {
-  import Vue from 'vue';
-
-  export default Vue;
-}
