--- webpack.renderer.config.js
@@ -1,17 +1,0 @@
-const rules = require('./webpack.rules');
-const plugins = require('./webpack.plugins');
-
-rules.push({
-  test: /\.css$/,
-  use: [{ loader: 'style-loader' }, { loader: 'css-loader' }],
-});
-
-module.exports = {
-  module: {
-    rules,
-  },
-  plugins: plugins,
-  resolve: {
-    extensions: ['.js', '.ts', '.jsx', '.tsx', '.css']
-  },
-};
